Rename `raspi::config` namespace to `raspi`
